change model's score db type

+++ db/db.py
@@ -59,7 +59,7 @@ models = sqlalchemy.Table(
     sqlalchemy.Column("type", sqlalchemy.String(length=250), nullable=False),
     sqlalchemy.Column("branch", sqlalchemy.String(length=100), nullable=False),
     sqlalchemy.Column("description", sqlalchemy.String(length=200), nullable=False),
-    sqlalchemy.Column("score", sqlalchemy.Float(), nullable=False),
+    sqlalchemy.Column("score", sqlalchemy.Numeric(), nullable=False),
 )
